Update default encoder test to pytest
