+++ requirements.txt
@@ -10,3 +10,5 @@ rerun-sdk
 pre-commit
 minio
 flow_vis
+pytest
+isort
