+++ tests/models/encoders/test_encoders.py
@@ -1,14 +1,10 @@
-"""
-Tests for all image encoders that force uniception functionalities
-"""
-
 import os
 import random
-import unittest
 from functools import lru_cache
 from typing import Tuple
 
 import numpy as np
+import pytest
 import requests
 import torch
 from PIL import Image
@@ -17,58 +13,55 @@ from uniception.models.encoders import *
 from uniception.models.encoders.image_normalizations import *
 
 
-class TestEncoders(unittest.TestCase):
-    def __init__(self, *args, **kwargs):
-        super(TestEncoders, self).__init__(*args, **kwargs)
+@pytest.fixture(scope="module")
+def norm_types():
+    return IMAGE_NORMALIZATION_DICT.keys()
 
-        self.norm_types = IMAGE_NORMALIZATION_DICT.keys()
 
-        # self.encoders = [
-        #     "croco",
-        #     "dust3r_224",
-        #     "dust3r_512",
-        #     "dust3r_512_dpt",
-        #     "mast3r_512",
-        #     "dinov2_large",
-        #     "dinov2_large_reg",
-        #     "dinov2_large_dav2",
-        #     "dinov2_giant",
-        #     "dinov2_giant_reg",
-        #     "radio_v2.5-b",
-        #     "radio_v2.5-l",
-        #     "e-radio_v2",
-        #     "naradio_v2.5-b",
-        #     "naradio_v2.5-l",
-        # ]
-        self.encoders = [
-            "naradio_v2.5-b",
-        ]
+@pytest.fixture(scope="module")
+def encoders():
+    return [
+        "croco",
+        "dust3r_224",
+        "dust3r_512",
+        "dust3r_512_dpt",
+        "mast3r_512",
+        "dinov2_base",
+        "dinov2_large",
+        "dinov2_large_reg",
+        "dinov2_large_dav2",
+        "dinov2_giant",
+        "dinov2_giant_reg",
+        "radio_v2.5-b",
+        "radio_v2.5-l",
+        "e-radio_v2",
+        "naradio_v2.5-b",
+        "naradio_v2.5-l",
+    ]
 
-        self.encoder_configs = [{}] * len(self.encoders)
 
-    def inference_encoder(self, encoder, input):
-        return encoder(input)
+@pytest.fixture(scope="module")
+def encoder_configs(encoders):
+    # Adjust the number of configs to match the number of encoders
+    return [{}] * len(encoders)
 
-    def test_make_dummy_encoder(self):
-        print("Testing Init of Dummy Encoder")
-        encoder = _make_encoder_test("dummy")
-        self.assertTrue(encoder is not None)
 
-    def test_all_encoder_basics(self):
-        for encoder, encoder_config in zip(self.encoders, self.encoder_configs):
-            print(f"Testing encoder: {encoder}")
+@pytest.fixture
+def device(request):
+    # Access the value of the custom option for device
+    device_str = request.config.getoption("--device")
+    if device_str == "gpu" and torch.cuda.is_available():
+        device = torch.device("cuda")  # Use the default CUDA device
+    else:
+        device = torch.device("cpu")
+    print(f"Using device: {device.type.upper()}")
+    return device
 
-            encoder = _make_encoder_test(encoder, **encoder_config)
-            self._check_baseclass_attribute(encoder)
-            self._check_norm_check_function(encoder)
 
-            if isinstance(encoder, UniCeptionViTEncoderBase):
-                self._check_vit_encoder_attribute(encoder)
-                self._test_vit_encoder_patch_size(encoder)
-
+@pytest.fixture
+def example_input(device):
     @lru_cache(maxsize=3)
     def _get_example_input(
-        self,
         image_size: Tuple[int, int],
         image_norm_type: str = "dummy",
         img_selection: int = 1,
@@ -82,80 +75,121 @@ class TestEncoders(unittest.TestCase):
         img = torch.from_numpy(np.array(image))
         viz_img = img.clone()
 
-        # Normalize the images
+        # Normalize the image
         image_normalization = IMAGE_NORMALIZATION_DICT[image_norm_type]
-
-        img_mean, img_std = image_normalization.mean, image_normalization.std
-
+        img_mean = image_normalization.mean
+        img_std = image_normalization.std
         img = (img.float() / 255.0 - img_mean) / img_std
 
-        # convert to BCHW format
-        img = img.permute(2, 0, 1).unsqueeze(0)
+        # Convert to BCHW format
+        img = img.permute(2, 0, 1).unsqueeze(0).to(device)
 
         if return_viz_img:
             return img, viz_img
         else:
             return img
 
-    def _test_vit_encoder_patch_size(self, encoder):
-        print(f"Testing patch size for encoder: {encoder.name}")
-        image_size = (14 * encoder.patch_size, 14 * encoder.patch_size)
-
-        img = self._get_example_input(image_size, encoder.data_norm_type)
-        # input and output of the encoder
-        encoder_input: ViTEncoderInput = ViTEncoderInput(
-            data_norm_type=encoder.data_norm_type,
-            image=img,
-        )
-
-        encoder_output = self.inference_encoder(encoder, encoder_input).features
-
-        self.assertTrue(isinstance(encoder_output, torch.Tensor))
-        self.assertTrue(encoder_output.shape[2] == 14)
-        self.assertTrue(encoder_output.shape[3] == 14)
-
-    def _check_baseclass_attribute(self, encoder):
-        self.assertTrue(hasattr(encoder, "name"))
-        self.assertTrue(hasattr(encoder, "size"))
-        self.assertTrue(hasattr(encoder, "data_norm_type"))
-
-        self.assertTrue(isinstance(encoder.name, str))
-        self.assertTrue(isinstance(encoder.size, str) or encoder.size is None)
-        self.assertTrue(isinstance(encoder.data_norm_type, str))
-
-        # Check if the data_norm_type is in the list of normalization types
-        self.assertTrue(encoder.data_norm_type in self.norm_types)
-
-    def _check_norm_check_function(self, encoder):
-        self.assertTrue(hasattr(encoder, "_check_data_normalization_type"))
-
-        encoder_notm_type = encoder.data_norm_type
-
-        try:
-            encoder._check_data_normalization_type(encoder_notm_type)
-        except AssertionError:
-            self.assertTrue(False)
-
-        try:
-            encoder._check_data_normalization_type("some_nonexistent_norm_type")
-            self.assertTrue(False)
-        except AssertionError:
-            pass
-
-    def _check_vit_encoder_attribute(self, encoder):
-        self.assertTrue(hasattr(encoder, "patch_size"))
-        self.assertTrue(isinstance(encoder.patch_size, int))
-        self.assertTrue(encoder.patch_size > 0)
+    return _get_example_input
 
 
-def seed_everything(seed=42):
-    """
-    Set the `seed` value for torch and numpy seeds. Also turns on
-    deterministic execution for cudnn.
+def inference_encoder(encoder, encoder_input):
+    # Encoder expects a ViTEncoderInput object
+    return encoder(encoder_input).features
 
-    Parameters:
-    - seed:     A hashable seed value
-    """
+
+def test_make_dummy_encoder(device):
+    print(f"Testing Init of Dummy Encoder on {device.type.upper()}")
+    encoder = _make_encoder_test("dummy").to(device)
+
+    # Check if the encoder has parameters
+    try:
+        params = list(encoder.parameters())
+        if not params:
+            print("Warning: The encoder has no parameters.")
+        else:
+            # Verify if the model is on the right device
+            assert params[0].is_cuda == (device.type == "cuda")
+
+    except Exception as e:
+        print(f"Error: {e}")
+        assert False  # Fail the test if any error occurs
+
+    assert encoder is not None
+
+
+def test_all_encoder_basics(encoders, encoder_configs, norm_types, example_input, encoder_name, device):
+    if encoder_name:
+        encoders = [encoder_name]  # Override default encoders with the one specified
+
+    for encoder_name, encoder_config in zip(encoders, encoder_configs):
+        print(f"Testing encoder: {encoder_name} on {device.type.upper()}")
+
+        encoder = _make_encoder_test(encoder_name, **encoder_config).to(device)
+        _check_baseclass_attribute(encoder, norm_types)
+        _check_norm_check_function(encoder)
+
+        if isinstance(encoder, UniCeptionViTEncoderBase):
+            _check_vit_encoder_attribute(encoder)
+            _test_vit_encoder_patch_size(encoder, example_input)
+
+
+def _check_baseclass_attribute(encoder, norm_types):
+    assert hasattr(encoder, "name")
+    assert hasattr(encoder, "size")
+    assert hasattr(encoder, "data_norm_type")
+
+    assert isinstance(encoder.name, str)
+    assert isinstance(encoder.size, str) or encoder.size is None
+    assert isinstance(encoder.data_norm_type, str)
+
+    # Check if the data_norm_type is in the list of normalization types
+    assert encoder.data_norm_type in norm_types
+
+
+def _check_norm_check_function(encoder):
+    assert hasattr(encoder, "_check_data_normalization_type")
+
+    encoder_notm_type = encoder.data_norm_type
+
+    try:
+        encoder._check_data_normalization_type(encoder_notm_type)
+    except AssertionError:
+        assert False
+
+    try:
+        encoder._check_data_normalization_type("some_nonexistent_norm_type")
+        assert False
+    except AssertionError:
+        pass
+
+
+def _check_vit_encoder_attribute(encoder):
+    assert hasattr(encoder, "patch_size")
+    assert isinstance(encoder.patch_size, int)
+    assert encoder.patch_size > 0
+
+
+def _test_vit_encoder_patch_size(encoder, example_input):
+    print(f"Testing {encoder.name} inference")
+    image_size = (14 * encoder.patch_size, 14 * encoder.patch_size)
+
+    img = example_input(image_size, encoder.data_norm_type)
+    # Create an instance of ViTEncoderInput with correct attributes
+    encoder_input = ViTEncoderInput(
+        data_norm_type=encoder.data_norm_type,
+        image=img,
+    )
+
+    encoder_output = inference_encoder(encoder, encoder_input)
+
+    assert isinstance(encoder_output, torch.Tensor)
+    assert encoder_output.shape[2] == 14
+    assert encoder_output.shape[3] == 14
+
+
+@pytest.fixture(scope="session", autouse=True)
+def seed_everything():
+    seed = 42
     random.seed(seed)
     os.environ["PYTHONHASHSEED"] = str(seed)
     np.random.seed(seed)
@@ -164,13 +198,5 @@ def seed_everything(seed=42):
     torch.backends.cudnn.benchmark = False
     print(f"Seed set to: {seed} (type: {type(seed)})")
 
-
-if __name__ == "__main__":
     # Turn XFormers off for testing on CPU
     os.environ["XFORMERS_DISABLED"] = "1"
-
-    # Seed everything for consistent testing
-    seed_everything()
-
-    # Test the Encoders
-    unittest.main()
