+++ uniception/models/utils/transformer_blocks.py
@@ -89,6 +89,7 @@ class Mlp(nn.Module):
 
 class Attention(nn.Module):
     "Self-Attention Layer"
+
     fused_attn: Final[bool]
 
     def __init__(
@@ -173,6 +174,7 @@ class Attention(nn.Module):
 
 class CrossAttention(nn.Module):
     "Cross-Attention Layer"
+
     fused_attn: Final[bool]
 
     def __init__(
